# fiat-payments-backend/src/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BrainSafes API",
    "description": "API completa para el ecosistema BrainSafes - Educación descentralizada con IA, NFTs y DeFi",
    "version": "1.0.0",
    "contact": {
      "name": "BrainSafes Team",
      "email": "support@brainsafes.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.brainsafes.com",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Webhook": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean"
          },
          "stats": {
            "type": "object",
            "properties": {
              "totalCalls": {
                "type": "number"
              },
              "successfulCalls": {
                "type": "number"
              },
              "failedCalls": {
                "type": "number"
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/webhooks": {
      "post": {
        "summary": "Registrar webhook",
        "description": "Registrar un nuevo webhook para recibir eventos",
        "tags": ["Webhooks"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url", "secret", "events"],
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "secret": {
                    "type": "string",
                    "minLength": 16
                  },
                  "events": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "minItems": 1
                  },
                  "options": {
                    "type": "object",
                    "properties": {
                      "timeout": {
                        "type": "number",
                        "minimum": 1000,
                        "maximum": 30000
                      },
                      "retries": {
                        "type": "number",
                        "minimum": 1,
                        "maximum": 10
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook registrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Webhook"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar webhooks",
        "description": "Obtener lista de webhooks registrados",
        "tags": ["Webhooks"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de webhooks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "totalWebhooks": {
                          "type": "number"
                        },
                        "activeWebhooks": {
                          "type": "number"
                        },
                        "webhooks": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Webhook"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
} 